backend/requirements.txt
# API
fastapi
uvicorn[standard]

# Base de Datos
sqlalchemy
psycopg2-binary

# Imágenes
Pillow

# IA Vision
groq

# Utilidades
python-multipart
python-dotenv
